+++ changes/CHANGES.rst
@@ -29,3 +29,4 @@ Unreleased
 - Initial Setup for Testing the CLI \(`#25 <https://github.com/Bibo-Joshi/chango/pull/25>`_\)
 - Initial Readme Content and Changelog Setup (`#26 <https://github.com/Bibo-Joshi/chango/pull/26>`_)
 - Create Change Notes from GitHub Events and Add GitHub Action (`#27 <https://github.com/Bibo-Joshi/chango/pull/27>`_)
+- Add Public `chango.config` Module (`#29 <https://github.com/Bibo-Joshi/chango/pull/29>`_)